+++ clients/shared_library/interfaces/update_course_phase_participation_status.ts
@@ -1,6 +1,6 @@
 import { PassStatus } from './course_phase_participation'
 
-export interface UpdateApplicationStatus {
+export interface UpdateCoursePhaseParticipationStatus {
   pass_status: PassStatus
   course_phase_participation_ids: string[]
 }
