+++ server/coursePhase/coursePhaseParticipation/coursePhaseParticipationDTO/update_course_phase_participation_status.go
@@ -1,11 +1,11 @@
-package applicationDTO
+package coursePhaseParticipationDTO
 
 import (
 	"github.com/google/uuid"
 	db "github.com/niclasheun/prompt2.0/db/sqlc"
 )
 
-type PutApplicationStatus struct {
+type UpdateCoursePhaseParticipationStatus struct {
 	PassStatus                  db.PassStatus `json:"pass_status"`
 	CoursePhaseParticipationIDs []uuid.UUID   `json:"course_phase_participation_ids"`
 }
